🔥 Simplify rewrap key modules (#391)
- Make it more clear that we expect the wrapping type to be a JWE crypto name, in this case `RSA-OAEP`
- TK soon: an update to the spec to clarify where and how to store the algorithm, and any parameters if we need them
