+++ .github/workflows/build.yaml
@@ -134,7 +134,7 @@ jobs:
       packages: read
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-node@v2
+      - uses: actions/setup-node@v3
         with:
           node-version: "18.x"
           registry-url: https://npm.pkg.github.com
