+++ containers/kas/kas_core/tdf3_kas_core/__init__.py
@@ -1,7 +1,5 @@
 """TDF3 KAS Core Exports."""
 
-from .models import Crypto  # noqa: F401
-
 from .models import Policy  # noqa: F401
 from .models import Entity  # noqa: F401
 from .models import KeyAccess  # noqa: F401
