+++ containers/kas/kas_core/tdf3_kas_core/models/__init__.py
@@ -8,8 +8,6 @@ from .key_master import KeyMaster  # noqa: F401
 
 from .context import Context  # noqa: F401
 
-from .crypto import Crypto  # noqa: F401
-
 from .attributes import AttributeValue  # noqa: F401
 from .attributes import AttributeSet  # noqa: F401
 from .attributes import DataAttributes  # noqa: F401
@@ -33,7 +31,7 @@ from .plugin_runner import RewrapPluginRunnerV2  # noqa: F401
 from .plugin_runner import UpsertPluginRunner  # noqa: F401
 from .plugin_runner import UpsertPluginRunnerV2  # noqa: F401
 
-from .wrapped_key import WrappedKey  # noqa: F401
+from .wrapped_keys import WrappedKey  # noqa: F401
 
 from .key_access import KeyAccess  # noqa: F401
