--- containers/kas/kas_core/tdf3_kas_core/models/crypto/__init__.py
@@ -1,3 +1,0 @@
-"""The Crypto model."""
-
-from .crypto import Crypto  # noqa: F401
