+++ containers/kas/kas_core/tdf3_kas_core/models/key_access/key_access_helpers_test.py
@@ -9,11 +9,11 @@ from pprint import pprint
 from cryptography.hazmat.primitives.ciphers.aead import AESGCM
 
 from tdf3_kas_core.models import WrappedKey
+from tdf3_kas_core.models.wrapped_keys import aes_encrypt_sha1
 
 from tdf3_kas_core.util import get_public_key_from_disk
 from tdf3_kas_core.util import get_private_key_from_disk
 from tdf3_kas_core.util import generate_hmac_digest
-from tdf3_kas_core.util import aes_encrypt_sha1
 from tdf3_kas_core.util import aes_gcm_encrypt
 
 from tdf3_kas_core.errors import KeyAccessError
