--- containers/kas/kas_core/tdf3_kas_core/models/wrapped_key/__init__.py
@@ -1,3 +1,0 @@
-"""WrappedKey models one key split."""
-
-from .wrapped_key import WrappedKey  # noqa :F401
