+++ containers/kas/kas_core/tdf3_kas_core/util/__init__.py
@@ -2,13 +2,8 @@
 
 # TODO why does this file exist, I'm pretty sure
 # python modules don't have to be like this
-from .keys import assure_public_key  # noqa: F401
-from .keys import assure_private_key  # noqa: F401
-
 from .cipher import aes_gcm_encrypt  # noqa: F401
 from .cipher import aes_gcm_decrypt  # noqa: F401
-from .cipher import aes_encrypt_sha1  # noqa: F401
-from .cipher import aes_decrypt_sha1  # noqa: F401
 
 from .keys import get_public_key_from_disk  # noqa: F401
 from .keys import get_private_key_from_disk  # noqa: F401
