+++ containers/kas/kas_core/tdf3_kas_core/util/cipher/__init__.py
@@ -2,6 +2,3 @@
 
 from .aes_gcm import aes_gcm_encrypt  # noqa: F401
 from .aes_gcm import aes_gcm_decrypt  # noqa: F401
-
-from .rsa_sha1 import aes_encrypt_sha1  # noqa: F401
-from .rsa_sha1 import aes_decrypt_sha1  # noqa: F401
