+++ containers/kas/kas_core/tdf3_kas_core/util/keys/__init__.py
@@ -11,6 +11,3 @@ from .get_keys_from_disk import get_symmetric_key_from_disk  # noqa: F401
 
 from .get_keys_from_pem import get_public_key_from_pem  # noqa: F401
 from .get_keys_from_pem import get_private_key_from_pem  # noqa: F401
-
-from .assure_key import assure_public_key  # noqa: F401
-from .assure_key import assure_private_key  # noqa: F401
